3. advanced/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.0.3
passlib[bcrypt]==1.7.4
python-jose==3.3.0
python-multipart==0.0.6
psycopg2-binary==2.9.9
alembic==1.12.1
email-validator==2.1.0.post1
httpx==0.25.2
pytest==7.4.3
